pyproject.toml
[tool.poetry]
name = "call-grafana-api"
version = "0.1.0"
description = "Python scripts to call Grafana APIs"
authors = ["Jijun Dai <jijun.dai@mindgeek.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
